+++ ompy/response.py
@@ -1,21 +1,25 @@
-# import sys, os
-# if sys.version_info[0] < 3:
-#     raise Exception("Must be using Python 3")
+# -*- coding: utf-8 -*-
+"""
+Implementation of a response matrix created from interpolation
+of experimental spectra. It takes the path to a folder (old format)
+or a zip file as its construction parameters.
+"""
+
 import os
+import zipfile as zf
 import numpy as np
 import pandas as pd
-# import matplotlib.pyplot as plt
-from scipy.interpolate import interp1d#, interp2d
+from pathlib import Path
+from typing import Union, Optional, Tuple
+from scipy.interpolate import interp1d
 import logging
 
-
-# from .firstgen import *
-# from .unfold import *
-from .rebin import *
-from .library import *
+from .rebin import rebin_1D
+from .library import div0
 from .decomposition import index
 from .gauss_smoothing import gauss_smoothing
 from .matrix import Matrix
+from .vector import Vector
 
 LOG = logging.getLogger(__name__)
 logging.captureWarnings(True)
@@ -23,71 +27,249 @@ logging.captureWarnings(True)
 DTYPE = np.float64
 
 
-# TODO 20190527: I just merged master into this old branch to start fixing the
-# response interpolation. I moved this file from oslo_method_python/ to ompy/.
-# The next step is to cythonize it properly to hopefully make it fast,
-# and implement it into the rest of the code structure, then test it.
-
-# j_test = 50
-
-
-def E_compton(Eg, theta):
+class Response():
     """
-    Calculates the energy of an electron that is scattered an angle
-    theta by a gamma-ray of energy Eg.
-    Adapted from MAMA, file "folding.f", which references
-    Canberra catalog ed.7, p.2.
-    Inputs:
-    Eg: Energy of gamma-ray in keV
-    theta: Angle of scatter in radians
-    Returns:
-    Energy Ee of scattered electron
-    """
-    # Return Eg if Eg <= 0.1, else use formula
-    # print("From E_compton(): Eg =", Eg, ", theta =", theta, ", formula =", Eg*Eg/511*(1-np.cos(theta)) / (1+Eg/511 * (1-np.cos(theta))))
-    return np.where(Eg > 0.1, Eg*Eg/511*(1-np.cos(theta)) / (1+Eg/511 * (1-np.cos(theta))), Eg)
-
-
-def corr(Eg, theta):
-    """
-    Function to correct number of counts due to delta(theta)
-    Adapted from MAMA in the file kelvin.f
-    It is dE/dtheta of the E(theta) in Eq. (2) in Guttormsen 1996.
-    """
-    return (Eg*Eg/511*np.sin(theta))/(1+Eg/511*(1-np.cos(theta)))**2
-
-
-def two_channel_split(E_centroid, E_array):
-    """
-    When E_centroid is between two bins in E_array, this function
-    returns the indices of the two nearest bins and the distance to
-    the lower bin. The distance to the higher bin is 1-floor_distance
-
-    Args:
-        E_centroid (double): The energy of the centroid (mid-bin)
-        E_array (np.array, double): The energy grid to distribute
+    The Resonse class that stores the data required to perform
+    the interpolation needed to create the response matrix
     """
 
-    a0 = E_array[0]
-    a1 = E_array[1]-E_array[0]
+    def __init__(self,
+                 path: Union[str, Path] = None):
+        """
+        The resonse object is initialized with the path to the source
+        files required to perform the interpolation. The path varaiable
+        can either be a folder (assuming "old" format) or a zip file
+        containing the files otherwise found in the folder in the "old"
+        format.
 
-    bin_as_float = (E_centroid - a0)/a1
-    i_floor = int(np.floor(bin_as_float))
-    i_ceil = int(np.ceil(bin_as_float))
-    floor_distance = (bin_as_float - i_floor)
+        TODO:
+            - adapt rutines for the possibility that not all cmp spectra have
+              the same binning
+        """
 
-    return i_floor, i_ceil, floor_distance
+        path = Path(path) if isinstance(path, str) else path
+
+        if path.is_dir():
+            self.resp, self.compton_matrix, self.Ecmp_array = self.LoadDir(
+                path)  # Better names would be adventagious
+        elif path.is_file():
+            self.resp, self.compton_matrix, self.Ecmp_array = self.LoadZip(
+                path)
+
+        # if compton was not smoothed before
+        self.smooth_compton = False
+
+        # after how many sigma to truncate gaussian smoothing
+        self.truncate = 6
+
+    def LoadZip(self,
+                path: Union[str, Path],
+                resp_name: Optional[str] = 'resp.csv',
+                spec_prefix: Optional[str] = 'cmp'):
+        """ Method for loading response file and compton spectra from zipfile.
+
+        Is assumes that path is a zip file that constains at least XX files.
+        At least one has to be a special summary table to be named `resp_name`.
+
+        Args:
+            path (Union[str, Path]): path to folder
+            resp_name (Optional[str], optional): name of file with
+                `response table`
+            spec_prefix (Optional[str], optional): Prefix for all spectra
+
+        Returns:
+            (tuple) containing
+                - **resp** (*DataFrame*): Information of the `response table`.
+                - **compton_matrix** (*ndarray*): matrix with compton counts.
+                    Shape is (N_incident, N_cmp)
+                - **last.E** (*ndarray*): energy array
+
+        """
+        path = Path(path) if isinstance(path, str) else path
+
+        zfile = zf.ZipFile(path, mode='r')
+
+        if not any(resp_name in name for name in zfile.namelist()):
+            raise ValueError('Response info not present')
+
+        resp = pd.read_csv(zfile.open(resp_name, 'r'))
+
+        # Verify that resp has all the required columns
+        if not set(['Eg', 'FWHM_rel_norm', 'Eff_tot', 'FE', 'SE', 'DE', 'c511']).issubset(resp.columns):
+            raise ValueError(f'{resp_name} missing one or more required columns')
+
+        # Verify that zip file contains a spectra for each energy
+        files = [spec_prefix + str(int(Eg)) for Eg in sorted(resp['Eg'])]
+        if not all(file in zfile.namelist() for file in files):
+            raise ValueError(f'One or more compton spectra is missing in {path}')
+
+        # Now we will read in all the Compton spectra
+        N_cmp = -1
+        a0_cmp, a1_cmp = -1, -1
+        # Get calibration and array length from highest-energy spectrum, because the spectra
+        # may have differing length but this is bound to be the longest.
+        with zfile.open(spec_prefix + str(int(max(resp['Eg']))), 'r') as file:
+            lines = file.readlines()
+            a0_cmp = float(str(lines[6]).split(",")[1])  # calibration
+            a1_cmp = float(str(lines[6]).split(",")[2])  # coefficients [keV]
+            N_cmp = int(lines[8][15:]) + 1  # 0 is first index
+
+        compton_matrix = np.zeros((len(resp['Eg']), N_cmp))
+        i = 0
+        for file in [zfile.open(file_name) for file_name in files]:
+            cmp_current = np.genfromtxt(file, comments="!")
+            compton_matrix[i, 0:len(cmp_current)] = cmp_current
+            i += 1
+
+        return resp, compton_matrix, np.linspace(a0_cmp, a1_cmp * (N_cmp - 1), N_cmp)
+
+    def LoadDir(self,
+                path: Union[str, Path],
+                resp_name: Optional[str] = 'resp.dat',
+                spec_prefix: Optional[str] = 'cmp'):
+        """
+        Method for loading response file and compton spectra from a folder.
+
+        Args:
+            path (Union[str, Path]): path to folder
+            resp_name (Optional[str], optional): name of file with
+                `response table`
+            spec_prefix (Optional[str], optional): Prefix for all spectra
+
+        Returns:
+            (tuple) containing
+                - **resp** (*DataFrame*): Information of the `response table`.
+                - **compton_matrix** (*ndarray*): matrix with compton counts.
+                    Shape is (N_incident, N_cmp)
+                - **last.E** (*ndarray*): energy array
+
+        Raises:
+            Exception: Description
+            NotImplementedError: Description
+        """
+
+        # Read resp.dat file, which gives information about the energy bins
+        # and discrete peaks
+        resp = []
+        Nlines = -1
+        with open(os.path.join(path, resp_name)) as file:
+            while True:
+                line = file.readline()
+                if not line:
+                    break
+                if line[0:22] == "# Next: Numer of Lines":
+                    # TODO: The above if test is hardly very robust. Find a
+                    # better solution.
+                    line = file.readline()
+                    Nlines = int(line)
+                    # print("Nlines =", Nlines)
+                    break
+
+            line = file.readline()
+            # print("line =", line)
+            if not line:
+                raise Exception("Error reading resp.dat")
+
+            for i in range(Nlines):
+                line = file.readline()
+                # print("line =", line)
+                row = np.array(line.split(), dtype="double")
+                resp.append(row)
+
+        # Unpack the resp matrix into its columns
+        resp = np.array(resp)
+        Eg_sim_array, fwhm_rel, Eff_tot, FE, SE, DE, c511 = resp.T
+        # "Eg_sim" means "gamma, simulated", and refers to the gamma energies
+        # where we have simulated Compton spectra.
 
 
-def interpolate_response(folderpath, Eout_array, fwhm_abs, return_table=False):
-    """ Interpolate response through "Fan"-mathod (Guttormsen1996)
+        # Get calibration and array length from highest-energy spectrum,
+        # because the spectra may have differing length,
+        # but the last is bound to be the longest.
+        N_Eg = len(Eg_sim_array)
+        fnames = []
+        for i in range(0, N_Eg):
+            fnames.append(f"{spec_prefix}{Eg_sim_array[i]:.0f}")
 
-    Assumes the source files are in the folder "folderpath",
-    and that they are formatted in a certain standard way.
-    The function interpolates the data to give a response matrix with
-    the desired energy binning specified by Eout_array.
+        last = Vector(path=os.path.join(path, fnames[-1]))
+        N_cmp = last.shape[0]
 
-    Args:
+        compton_matrix = np.zeros((N_Eg, N_cmp))
+        # Read the rest:
+        for i in range(0, N_Eg):
+            cmp_current = Vector(path=os.path.join(path, fnames[i]))
+            if cmp_current.calibration() != last.calibration():
+                raise NotImplementedError("Currently, all cmp calibrations"
+                                          "have to be the same")
+            compton_matrix[i, 0:len(cmp_current.values)] = cmp_current.values
+
+        resp = pd.DataFrame(data={
+            'Eg': Eg_sim_array,
+            'FWHM_rel_norm': fwhm_rel,
+            'Eff_tot': Eff_tot,
+            'FE': FE,
+            'SE': SE,
+            'DE': DE,
+            'c511': c511})
+        return resp, compton_matrix, last.E
+
+    def get_probabilities(self):
+        # total number of counts for each of the loaded responses
+        self.sum_spec = self.compton_matrix.sum(axis=1) \
+            + self.resp['FE'] + self.resp['SE'] + self.resp['DE'] \
+            + self.resp['c511']
+        self.sum_spec = np.array(self.sum_spec)
+
+        # normalize "compton" spectra
+        self.cmp_matrix = div0(self.compton_matrix,
+                               self.sum_spec.reshape((len(self.sum_spec), 1)))
+        # Vector of total Compton probability
+        self.pcmp = self.cmp_matrix.sum(axis=1)
+
+        # Full energy, single escape, etc:
+        self.pFE = div0(self.resp['FE'], self.sum_spec)
+        self.pSE = div0(self.resp['SE'], self.sum_spec)
+        self.pDE = div0(self.resp['DE'], self.sum_spec)
+        self.p511 = div0(self.resp['c511'], self.sum_spec)
+
+        # Interpolate the peak structures except Compton (handled separately)
+        def interpolate(y, fill_value="extrapolate"):
+            return interp1d(self.resp['Eg'], y,
+                            kind="linear", bounds_error=False,
+                            fill_value=fill_value)
+
+        self.f_pcmp = interpolate(self.pcmp)
+        self.f_pFE = interpolate(self.pFE)
+        self.f_pSE = interpolate(self.pSE, fill_value=0)
+        self.f_pDE = interpolate(self.pDE, fill_value=0)
+        self.f_p511 = interpolate(self.p511, fill_value=0)
+        self.f_fwhm_rel_perCent_norm = interpolate(self.resp['FWHM_rel_norm'])
+        # TODO: Should this be extrapolated, too?
+        self.f_Eff_tot = interpolate(self.resp['Eff_tot'])
+
+        fwhm_rel_1330 = (self.fwhm_abs / 1330 * 100)
+        self.f_fwhm_rel_perCent = interpolate(self.resp['FWHM_rel_norm']
+                                              * fwhm_rel_1330)
+        def f_fwhm_abs(E):
+            return E * self.f_fwhm_rel_perCent(E)/100
+
+        self.f_fwhm_abs = f_fwhm_abs
+
+    def interpolate(self,
+                    Eout: np.ndarray = None,
+                    fwhm_abs: float = None,
+                    return_table: bool = False):
+        """ Interpolated the response matrix
+
+        Perform the interpolation for the energy range specified in Eout with
+        FWHM at 1332 keV given by FWHM_abs (in keV).
+
+        The interpolation is split into energy regions. Below the
+        back-scattering energy Ebsc we interpolate linearly, then we apply the
+        "fan method" (Guttormsen 1996) in the region from Ebsc up to the
+        Compton edge, then linear extrapolation again the rest of the way.
+
+        Args:
         folderpath: The path to the folder containing Compton spectra and resp.dat
         Eout_array: The desired energies of the output response matrix.
         fwhm_abs: The experimental absolute full-width-half-max at 1.33 MeV.
@@ -95,423 +277,374 @@ def interpolate_response(folderpath, Eout_array, fwhm_abs, return_table=False):
                   real FWHM for unfolding.
         return_table (optional): Returns "all" output, see below
 
-    Returns:
-        response (Matrix): Response matrix with incident energy on the "Ex"
+        Returns:
+         - response (Matrix): Response matrix with incident energy on the "Ex"
                            axis and the spectral response on the "Eg" axis
-        response_table (Dataframe)
+         - response_table (optinal, Dataframe): Table with efficiencies,
+                FE, SE (...) probabilities, and so on
+        """
+        self.Eout = Eout
+        self.fwhm_abs = fwhm_abs
 
-    """
-    if return_table not in (True, False):
-        raise ValueError("return_table must be a bool,now {}".format(return_table))
+        self.get_probabilities()
+        self.iterpolate_checks()
 
-    assert(1e-1 <= fwhm_abs <= 100), "Check the fwhm_abs, probably it's wrong."\
-        "\nNormal Oscar≃30, Now: {}".format(fwhm_abs)
+        N_out = len(Eout)
+        self.N_out = N_out
+        fwhm_abs_array = Eout * self.f_fwhm_rel_perCent(Eout) / 100
 
-    # Define helping variables from input
-    N_out = len(Eout_array)
-    a0_out, a1_out = Eout_array[0], Eout_array[1]-Eout_array[0]
-    # print("a0_out, a1_out =", a0_out, a1_out)
-    # cdef int i
+        R = np.zeros((N_out, N_out))
+        # Loop over rows of the response matrix
+        # TODO for speedup: Change this to a cython
+        for j, E in enumerate(Eout):
 
-    # Read resp.dat file, which gives information about the energy bins
-    # and discrete peaks
-    resp = []
-    Nlines = -1
-    with open(os.path.join(folderpath, "resp.dat")) as file:
-        while True:
-            line = file.readline()
-            if not line:
-                break
-            if line[0:22] == "# Next: Numer of Lines":
-                # TODO: The above if test is hardly very robust. Find a better solution.
-                line = file.readline()
-                Nlines = int(line)
-                # print("Nlines =", Nlines)
-                break
+            # Find maximal energy for current response (+n*sigma) function,
+            # -> Better if the lowest energies of the simulated spectra are
+            #    above the gamma energy to be extrapolated
+            oneSigma = fwhm_abs * self.f_fwhm_rel_perCent_norm(E) / 2.35
+            Egmax = E + 6 * oneSigma
+            i_Egmax = min(index(Eout, Egmax), N_out-1)
+            # LOG.debug("Response for E: {E:.0f} calc. up to {Egmax:.0f}")
 
-        line = file.readline()
-        # print("line =", line)
-        if not line:
-            raise Exception("Error reading resp.dat")
+            compton = self.get_closest_compton(E)
 
-        for i in range(Nlines):
-            line = file.readline()
-            # print("line =", line)
-            row = np.array(line.split(), dtype="double")
-            resp.append(row)
+            R_low, i_bsc = self.linear_backscatter(E, compton)
+            R_fan, i_last = self.fan_method(E, compton,
+                                            i_start=i_bsc+1, i_stop=i_Egmax)
+            if i_last == i_bsc+1:  # fan method didn't do anything
+                i_last -= 1
+            R_high = self.linear_to_end(E, compton,
+                                        i_start=i_last+1, i_stop=i_Egmax)
+            R[j, :] += R_low + R_fan + R_high
 
-    # Unpack the resp matrix into its columns
-    resp = np.array(resp)
-    Eg_sim_array, fwhm_rel, Eff_tot, FE, SE, DE, c511 = resp.T
-    a0_sim, a1_sim = Eg_sim_array[0], Eg_sim_array[1]-Eg_sim_array[0]
-    # print("a0_sim, a1_sim =", a0_sim, a1_sim, flush=True)
-    # "Eg_sim" means "gamma, simulated", and refers to the gamma energies where we have simulated Compton spectra.
+            # coorecton below E_sim[0]
+            if E < self.resp['Eg'][0]:
+                R[j, j + 1:] = 0
 
-    if Eout_array.min() < Eg_sim_array.min():
-        LOG.info("Note: The response below {:.0f} keV".format(Eg_sim_array.min())\
-            + "is interpolation only, as there are no simulations available.")
+            discrete_peaks = self.discrete_peaks(j, fwhm_abs_array)
+            R[j, :] += discrete_peaks
 
-    if Eout_array.max() > Eg_sim_array.max():
-        # Actually I don't know why we shouldn't be able to
-        # interpolate here, too
-        raise ValueError("Maximum energy cannot be larger than largest "
-                         "simulated response function, which is "
-                         "E_max = {:f}".format(Eg_sim_array.max())
-                         )
+            # smooth if compton background was not smoothed before (?)
+            # if performed here, no need to smooth twice, see discrete_peaks
+            if self.smooth_compton:
+                R[j, :] = gauss_smoothing(R[j, :], self.Eout,
+                                          fwhm_abs_array,
+                                          truncate=self.truncate)
 
+            # normalize (preserve probability)
+            R[j, :] = div0(R[j, :], np.sum(R[j, :]))
 
-    # Read in Compton spectra for each Eg channel:
-    N_Eg = len(Eg_sim_array)
-    # Read first Compton spectrum to get number of energy channels in each:
-    N_cmp = -1
-    a0_cmp, a1_cmp = -1, -1
-    # Get calibration and array length from highest-energy spectrum, because the spectra
-    # may have differing length but this is bound to be the longest.
-    with open(os.path.join(folderpath,"cmp"+str(int(Eg_sim_array[-1])))) as file:
-        lines = file.readlines()
-        a0_cmp = float(lines[6].split(",")[1]) # calibration
-        a1_cmp = float(lines[6].split(",")[2]) # coefficients [keV]
-        N_cmp = int(lines[8][15:]) +1 # 0 is first index
-    # print("a0_cmp, a1_cmp, N_cmp = ", a0_cmp, a1_cmp, N_cmp)
-    compton_matrix = np.zeros((N_Eg, N_cmp))
-    # Read the rest:
-    for i in range(0,N_Eg):
-        fn = "cmp"+str(Eg_sim_array[i])
-        cmp_current = np.genfromtxt(os.path.join(folderpath,"cmp"+str(int(Eg_sim_array[i]))), comments="!")
-        compton_matrix[i,0:len(cmp_current)] = cmp_current
+        # Remove any negative elements from response matrix:
+        if len(R[R < 0]) != 0:
+            LOG.debug(f"{len(R[R < 0])} entries in R were set to 0")
+            R[R < 0] = 0
 
-    # print("compton_matrix =", compton_matrix)
-    # compton_matrix = np.tril(compton_matrix)
+        response = Matrix(values=R, Eg=Eout, Ex=Eout)
 
-    # Normalize total spectrum to 1, including FE, SE, etc:
-    # Make a vector containing the sums for each step along the Eg_sim dimension:
-    sum_spec = compton_matrix.sum(axis=1) + FE + SE + DE + c511
+        if return_table:
+            # Return the response matrix, as well as the other structures, FWHM and efficiency, interpolated to the Eout_array
+            response_table = {'E': Eout,
+                              'fwhm_abs': fwhm_abs_array,
+                              'fwhm_rel_%': self.f_fwhm_rel_perCent(Eout),
+                              'fwhm_rel': self.f_fwhm_rel_perCent(Eout)/100,
+                              'eff_tot': self.f_Eff_tot(Eout),
+                              'pcmp': self.f_pcmp(Eout),
+                              'pFE': self.f_pFE(Eout),
+                              'pSE': self.f_pSE(Eout),
+                              'pDE': self.f_pDE(Eout),
+                              'p511': self.f_p511(Eout)}
+            response_table = pd.DataFrame(data=response_table)
+            return response, response_table
+        else:
+            return response
 
-    # Normalize each part of each spectrum:
-    cmp_matrix = div0(compton_matrix,sum_spec.reshape((len(sum_spec),1))) # Still a matrix with Eg_sim along first axis, Ecmp along second
-    pcmp = cmp_matrix.sum(axis=1) # Vector of total Compton probability
-    # Full energy, single escape, etc:
-    pFE = div0(FE,sum_spec)
-    pSE = div0(SE,sum_spec)
-    pDE = div0(DE,sum_spec)
-    p511 = div0(c511,sum_spec)
+    def iterpolate_checks(self):
+        """ Check on the inputs to `interpolate` """
+        assert(1e-1 <= self.fwhm_abs <= 1000), \
+            "Check the fwhm_abs, probably it's wrong."\
+            "\nNormal Oscar≃30 keV, Now: {} keV".format(self.fwhm_abs)
 
+        Eout = self.Eout
+        if len(Eout) <= 1:
+            raise ValueError("Eout should have more elements than 1" \
+                             f"now {len(Eout)}")
 
-    # == Interpolate the peak structures except Compton, which is handled separately ==
-    f_pcmp = interp1d(Eg_sim_array, pcmp, kind="linear", bounds_error=False, fill_value="extrapolate")
-    f_pFE = interp1d(Eg_sim_array, pFE, kind="linear", bounds_error=False, fill_value="extrapolate")
-    f_pSE = interp1d(Eg_sim_array, pSE, kind="linear", bounds_error=False, fill_value=0)
-    f_pDE = interp1d(Eg_sim_array, pDE, kind="linear", bounds_error=False, fill_value=0)
-    f_p511 = interp1d(Eg_sim_array, p511, kind="linear", bounds_error=False, fill_value=0)
-    f_fwhm_rel_perCent_norm = interp1d(Eg_sim_array, fwhm_rel, kind="linear", bounds_error=False, fill_value="extrapolate")
-    f_Eff_tot = interp1d(Eg_sim_array, Eff_tot, kind="linear", bounds_error=False, fill_value=0)
+        assert abs(self.f_fwhm_rel_perCent_norm(1330) - 1) < 0.05, \
+            "Response function format not as expected." \
+            "In the Mama-format, the 'f_fwhm_rel_perCent' column denotes"\
+            "the relative fwhm (= fwhm/E), but normalized to 1 at 1.33 MeV."\
+            f"Now it is: {self.f_fwhm_rel_perCent_norm(1330)} at 1.33 MeV."
 
-    assert abs(f_fwhm_rel_perCent_norm(1330)-1) < 0.05, \
-        "Response function format not as expected. In the Mama-format, the"\
-        "'f_fwhm_rel_perCent' column denotes the relative fwhm (= fwhm/E)," \
-        "but normalized to 1 at 1.33 MeV."\
-        "Now it is: {} at 1.33 MeV.".format(f_fwhm_rel_perCent_norm(1330))
-    fwhm_rel_1330 = (fwhm_abs/1330*100)
-    f_fwhm_rel_perCent = interp1d(Eg_sim_array, fwhm_rel*fwhm_rel_1330,
-                                  kind="linear",
-                                  bounds_error=False,
-                                  fill_value="extrapolate")
+    def get_closest_compton(self, E: float) -> Tuple[int, int]:
+        """Find and rebin closest energies from available response functions
 
-    fwhm_abs_array = Eout_array*f_fwhm_rel_perCent(Eout_array)/100
+        If `E < self.resp['Eg'].min()` the compton matrix will be replaced
+        by an array of zeros.
 
-    # DEBUG:
-    # print("p511 vector from resp.dat =", p511)
-    # END DEBUG
+        Args:
+            E (float): Description
 
+        Returns:
+            Tuple[int, int]: `ilow` and `ihigh` are indices of closest energies
+        """
+        N = len(self.resp['Eg'])
+        # ilow = 0
+        ihigh = np.searchsorted(self.resp['Eg'], E, side="right")
+        if ihigh == N:  # E > self.resp['Eg'].max()
+            ihigh -= 1
+        ilow = ihigh-1
 
+        # Select the Compton spectra, called Fs1 and Fs2 in MAMA.
+        Ehigh = self.resp['Eg'][ihigh]
+        cmp_high = self.cmp_matrix[ihigh, :]
+        if ilow < 0:  # E < self.resp['Eg'].min()
+            Elow = 0
+            cmp_low = np.zeros_like(cmp_high)
+        else:
+            Elow = self.resp['Eg'][ilow]
+            cmp_low = self.cmp_matrix[ilow, :]
 
-    # == Make response matrix by interpolating to Eg_sim_array ==
-    Ecmp_array = np.linspace(a0_cmp, a1_cmp*(N_cmp-1), N_cmp)
-    # print("Ecmp_array =", Ecmp_array)
+        cmp_low = rebin_1D(cmp_low, self.Ecmp_array, self.Eout)
+        cmp_high = rebin_1D(cmp_high, self.Ecmp_array, self.Eout)
 
-    # TODO normalize once and for all outside j loop
-    # f, ax = plt.subplots(1,1)
-    # for i_plt in [0,5,10,20,50]:
-        # ax.plot(Ecmp_array, compton_matrix[i_plt,:]/compton_matrix[i_plt,:].sum(), label="Eg = {:.0f}".format(Eg_sim_array[i_plt]))
+        compton = {"ilow": ilow,
+                   "ihigh": ihigh,
+                   "Elow": Elow,
+                   "Ehigh": Ehigh,
+                   "counts_low": cmp_low,
+                   "counts_high": cmp_high}
 
-    # We need to use the interpolation scheme given in Guttormsen 1996.
-    # Brute-force it with loops to make sure I get it right (based on MAMA code)
-    # After all, this is done once and for all, so it does not need to be lightning fast
+        return compton
 
-    # Start looping over the rows of the response function,
-    # indexed by j to match MAMA code:
-    # Egmin = 30 # keV -- this is universal (TODO: Is it needed?)
-    Egmin = Eout_array[0]
-    i_Egmin = index(Eout_array, Egmin)
+    def linear_cmp_interpolation(self, E: float, compton: dict,
+                                 fill_value: str = "extrapolate")-> np.ndarray:
+        """Linear interpolation between the compton spectra
 
-    # Allocate response matrix array:
-    R = np.zeros((N_out, N_out))
-    # Loop over rows of the response matrix
-    # TODO for speedup: Change this to a cython .pyx, declare the j variable.
-    #                   + run a Cython profiler, probably use memory views and
-    #                   other tweaks to speedup (see rebin.pyx for examples).
-    for j in range(N_out):
-        E_j = Eout_array[j]
-        # Skip if below lower threshold
-        if E_j < Egmin:
-            continue
+        Args:
+            E (float): Incident energy
+            compton (dict): Dict. with information about the compton spectra
+               to interpolate between
+            fill_value (str, optional): Fill value beyond boundaries
 
-        # Find maximal energy for current response function,
-        # Changed to 1*sigma, or whatever this means
-        # -> Better if the lowest energies of the simulated spectra are above
-        # the gamma energy to be extrapolatedu
-        Egmax = E_j + 1*fwhm_abs*f_fwhm_rel_perCent_norm(E_j)/2.35 #FWHM_rel.max()/2.35
-        i_Egmax = min(index(Eout_array, Egmax), N_out)
-        # print("i_Egmax =", i_Egmax)
+        Returns:
+            f_cmp (nd.array): Interpolated values
+        """
+        x = np.array([compton["Elow"], compton["Ehigh"]])
+        y = np.vstack([compton["counts_low"], compton["counts_high"]])
+        f_cmp = interp1d(x, y, kind="linear", bounds_error=False,
+                         fill_value=fill_value,
+                         axis=0)
+        return f_cmp(E)
 
-        # MAMA unfolds with 1/10 of real FWHM for convergence reasons.
-        # But let's stick to letting FWHM denote the actual value, and divide by 10 in computations if necessary.
+    def linear_backscatter(self, E: float,
+                           compton: dict) -> Tuple[np.ndarray, int]:
+        """Interpolate one-to-one up to the backscatter peak
 
-        # Find the closest energies among the available response functions, to interpolate between:
-        i_g_sim_low = 0
-        try:
-            i_g_sim_low = np.where(Eg_sim_array <= E_j)[0][-1]
-        except IndexError:
-            pass
-        i_g_sim_high = N_Eg
-        try:
-            i_g_sim_high = np.where(Eg_sim_array >= E_j)[0][0]
-        except IndexError:
-            pass
-        # When E_out[j] is below lowest Eg_sim_array element? Interpolate between two larger?
-        if i_g_sim_low == i_g_sim_high:
-            if i_g_sim_low > 0:
-                i_g_sim_low -= 1
-            else:
-                i_g_sim_high += 1
+        Args:
+            E (float): Incident energy
+            compton (dict): Dict. with information about the compton spectra
+               to interpolate between
 
-        Eg_low = Eg_sim_array[i_g_sim_low]
-        Eg_high = Eg_sim_array[i_g_sim_high]
+        Returns:
+            Tuple[np.ndarray, int]: `R` is Response for `E`, and
+                `i_bc` is index of backscatter peak
 
-        # Next, select the Compton spectra at index i_g_sim_low and i_g_sim_high. These are called Fs1 and Fs2 in MAMA.
-        # print("Eg_low =", Eg_low, "Eg_high =", Eg_high)
-        # print("i_g_sim_low =", i_g_sim_low, "i_g_sim_high =", i_g_sim_high, flush=True)
+        """
+        Eedge = self.E_compton(E, theta=np.pi)  # compton-edge energy
+        Ebsc = E - Eedge  # back-scattering energy
+        i_bsc = index(self.Eout, Ebsc)
+        R = np.zeros(self.N_out)
 
-        cmp_low = cmp_matrix[i_g_sim_low,:]
-        cmp_high = cmp_matrix[i_g_sim_high,:]
-        # These need to be recalibrated from Ecmp_array to Eout_array:
-        cmp_low = rebin_1D(cmp_low, Ecmp_array, Eout_array)
-        cmp_high = rebin_1D(cmp_high, Ecmp_array, Eout_array)
-        # print("Eout_array[{:d}] = {:.1f}".format(j, E_j), "Eg_low =", Eg_sim_array[i_g_sim_low], "Eg_high =", Eg_sim_array[i_g_sim_high], flush=True)
+        fcmp = self.linear_cmp_interpolation(E, compton)
+        R[:i_bsc+1] = fcmp[:i_bsc+1]
+        R[R < 0] = 0
+        return R, i_bsc
 
-        # The interpolation is split into energy regions.
-        # Below the back-scattering energy Ebsc we interpolate linearly,
-        # then we apply the "fan method" (Guttormsen 1996) in the region
-        # from Ebsc up to the Compton edge, then linear extrapolation again the rest of the way.
+    def linear_to_end(self, E: float, compton: dict,
+                      i_start: int, i_stop: int) -> np.ndarray:
+        """Interpolate one-to-one from the last fan energy to the Emax
 
-        # Get maximal energy by taking 6*sigma above full-energy peak
-        E_low_max = Eg_low + 6*fwhm_abs_array[i_g_sim_low]/2.35
-        i_low_max = min(index(Eout_array, E_low_max), N_out-1)
-        E_high_max = Eg_high + 6*fwhm_abs_array[i_g_sim_high]/2.35
-        i_high_max =min(index(Eout_array, E_high_max), N_out-1)
-        # print("E_low_max =", E_low_max, "E_high_max =", E_high_max, flush=True)
+        Args:
+            E (float): Incident energy
+            compton (dict): Dict. with information about the compton spectra
+               to interpolate between
+            i_start (int): Index where to start (usually end of fan method)
+            i_stop (int): Index where to stop (usually E+n*resolution)
 
-        # Find back-scattering Ebsc and compton-edge Ece energy of the current Eout energy:
-        Ece = E_compton(E_j, theta=np.pi)
-        Ebsc = E_j - Ece
-        # if E_j==200:
-        #     print(E_j)
-        #     print("Ece =", Ece)
-        #     print("Ebsc =", Ebsc)
-        # Indices in Eout calibration corresponding to these energies:
+        Returns:
+            np.ndarray: Response for `E`
+        """
+        R = np.zeros(self.N_out)
+        fcmp = self.linear_cmp_interpolation(E, compton)
+        R[i_start:i_stop+1] = fcmp[i_start:i_stop+1]
+        R[R < 0] = 0
+        return R
 
-        i_ce_out = min(index(Eout_array, Ece), i_Egmax)
-        i_bsc_out = min(index(Eout_array, Ebsc), i_Egmin)
+    def fan_method(self, E: float, compton: dict,
+                   i_start: int, i_stop: int) -> Tuple[np.ndarray, int]:
+        """Fan method
 
-        # print("i_ce_out =", i_ce_out, ", i_bsc_out =", i_bsc_out, ", i_Egmax =", i_Egmax)
+        Args:
+            E (float): Incident energy
+            compton (dict): Dict. with information about the compton spectra
+               to interpolate between
+            i_start (int): Index where to start (usually end of backscatter)
+            i_stop (int): Index where to stop (usually E+n*resolution). Note
+                that it can be stopped earlier, which will be reported through
+                `i_last`
 
+        Returns:
+            Tuple[np.ndarray, int]: `R` is Response for `E`, and `i_last` last
+                index of fan-method
+        """
+        R = np.zeros(self.N_out)
 
-        # ax.axvline(Ebsc)
-        # ax.axvline(Ece)
+        Ece = self.E_compton(E, theta=np.pi)
+        i_E_max = min(i_stop, self.N_out)
+        i_ce_max = min(index(self.Eout, Ece), i_E_max)
 
+        Esim_low = compton["Elow"]
+        Esim_high = compton["Ehigh"]
 
-        # Interpolate one-to-one up to j_bsc_out:
+        def lin_interpolation(x, x0, y0, x1, y1):
+            return y0 + (y1-y0)*(x-x0)/(x1-x0)
 
-        for i in range(0,i_bsc_out):
-            R[j,i] = cmp_low[i] + (cmp_high[i]-cmp_low[i])*(E_j - Eg_low)/(Eg_high-Eg_low)
-            if R[j,i] < 0:
-                # print("R[{:d},{:d}] = {:.2f}".format(j,i,R[j,i]), flush=True)
-                R[j,i] = 0 # TODO make this faster by indexing at the end
+        i_last = i_start  # Keep track of how far up the fan method goes
+        for i in range(i_start, i_ce_max):
+            # In Mama: E -> Egam, Ei -> E [Fabio]
+            # Energy of current point in interpolated spectrum
+            Ei = self.Eout[i]
+            if Ei < 0.1 or Ei > Ece:
+                continue
+            z = div0(Ei, (E / 511 * (E - Ei)))
+            theta = np.arccos(1 - z)
+            if theta > 0 and theta < np.pi:
+                # Determine interpolation indices in low and high arrays
+                # by Compton formula
+                Ecmp_ = self.E_compton(Esim_low, theta)
+                i_low_interp = index(self.Eout, Ecmp_)
+                Ecmp_ = self.E_compton(Esim_high, theta)
+                i_high_interp = index(self.Eout, Ecmp_)
 
+                c1 = compton["counts_low"][i_low_interp]
+                c2 = compton["counts_high"][i_high_interp]
 
+                # apply correction
+                c1 *= self.dE_dtheta(Esim_low, theta)
+                c2 *= self.dE_dtheta(Esim_high, theta)
 
-        # Then interpolate with the fan method up to j_ce_out:
-        z = 0 # Initialize variable
-        i_last = i_bsc_out # Keep track of how far up the fan method goes
-        i_low_last = i_bsc_out
-        i_high_last = i_bsc_out
+                # essential equation c(E), which is below (2)
+                interpol = lin_interpolation(E, Esim_low, c1, Esim_high, c2)
+                R[i] = interpol / self.dE_dtheta(E, theta)
+                i_last = i
 
-        for i in range(i_bsc_out, i_ce_out):
-            E_i = Eout_array[i] # Energy of current point in interpolated spectrum
-            if E_i > 0.1 and E_i < Ece:
-                if np.abs(E_j - E_i) > 0.001:
-                    z = E_i/(E_j/511 * (E_j - E_i))
-                theta = np.arccos(1-z)
-                # print("theta = ", theta, flush=True)
-                if theta > 0 and theta < np.pi:
-                    # Determine interpolation indices in low and high arrays
-                    # by Compton formula
-                    Ecmp_ = E_compton(Eg_low, theta)
-                    i_low_interp = min(index(Eout_array, Ecmp_), i_bsc_out)
-                    Ecmp_ = E_compton(Eg_high, theta)
-                    i_high_interp = min(index(Eout_array, Ecmp_), i_high_max)
-                    FA = (cmp_high[i_high_interp]*corr(Eg_high, theta)
-                          - cmp_low[i_low_interp]*corr(Eg_low, theta))
-                    FB = cmp_low[i_low_interp]*corr(Eg_low, theta) + FA*(E_j - Eg_low)/(Eg_high - Eg_low)
-                    R[j, i] = FB/corr(E_j, theta)
-                    i_last = i
-                    i_low_last = i_low_interp
-                    i_high_last = i_high_interp
+        if len(R[R < 0]) != 0:
+            LOG.debug(f"In fan method, {len(R[R < 0])} entries in R"
+                      "are negative and now set to 0")
+            R[R < 0] = 0
 
+        return R, i_last
 
-        # Interpolate 1-to-1 the last distance up to E+6*sigma
-        # print("i_Egmax =", i_Egmax, "Egmax =", Egmax, ", i_last =", i_last, flush=True)
-        # Check if this is needed:
-        if i_last >= i_Egmax:
-            continue
-        s_low = (i_low_max-i_low_last)/(i_Egmax-i_last)
-        s_high = (i_high_max-i_high_last)/(i_Egmax-i_last)
-
-        for i in range(i_last, i_Egmax):
-            i_low_interp = min(int(i_low_last + s_low*(i-i_last) + 0.5), i_low_max)
-            i_high_interp = min(int(i_high_last + s_high*(i-i_last) + 0.5), i_high_max)
-            R[j,i] = cmp_low[i_low_interp] + (cmp_high[i_high_interp]-cmp_low[i_low_interp])*(E_j-Eg_low)/(Eg_high-Eg_low)
-            # print("Last bit of interpolation: R[{:d},{:d}] = {:.2f}".format(j,i,R[j,i]), flush=True)
-            # if R[j,i] < 0:
-            #     print("R[j,i] =", R[j,i], flush=True)
-            #     R[j,i] = 0
-
-        # coorecton below E_sim[0]
-        if E_j < Eg_sim_array[0]:
-            R[j,j+1:]=0
-
-        # DEBUG: Plot cmp_low and cmp_high:
-        # if 50 < E_j <= 55 or 200 < E_j <= 205 or 500 < E_j <= 505:
-        #     if 50 < E_j <= 55: fig, ax = plt.subplots()
-        #     # ax.plot(Eout_array, cmp_low, label="cmp_low")
-        #     # ax.plot(Eout_array, cmp_high, label="cmp_high")
-        #     ax.plot(Eout_array, R[j, :], label="R[j, :]")
-        #     ax.plot(Eout_array, R1, "--",label="R[j, :]")
-        #     # ax.plot(Eout_array, cmp_high, label="cmp_high")
-        #     plt.show()
-
-        # Note: We choose not to smoothe the Compton spectrum, because the
-        # simulated Compton spectra stored in file are smoothed already.
-        # To apply smoothing to the Compton spectrum, you could do something like
-        # R[j, :] = gauss_smoothing(R[j, :], Eout_array,
-        #                           fwhm_abs_array)
-
-
-
-
-        # === Add peak structures to the spectrum: ===
-        discrete_peaks = np.zeros(N_out)
-        E_fe = Eout_array[j] + a1_out/2  # Evaluate energies in middle-bin
+    def discrete_peaks(self, i_response, fwhm_abs_array):
+        discrete_peaks = np.zeros(self.N_out)
+        Eout = self.Eout
+        E_fe = Eout[i_response]
 
         # Add full-energy peak, which should be at energy corresponding to
-        # index j:
-        # full_energy = np.zeros(N_out)  # Allocate with zeros everywhere
-        # full_energy[j] = f_pFE(E_fe)  # Full probability into sharp peak
-        discrete_peaks[j] = f_pFE(E_fe)
+        # index i_response:
+        discrete_peaks[i_response] = self.f_pFE(E_fe)
 
-        # Smoothe it:
-        # full_energy = gauss_smoothing(full_energy, Eout_array,
-                                      # fwhm_abs_array)
-        # R[j, :] += full_energy
-
-        # Add single-escape peak, at index i_se
+        # Add single-escape peak
         E_se = E_fe - 511
-        if E_se >= 0 and E_se >= Eout_array[0]:
+        if E_se >= 0 and E_se >= Eout[0]:
             i_floor, i_ceil, floor_distance\
-                = two_channel_split(E_se, Eout_array)
-            # single_escape = np.zeros(N_out)  # Allocate with zeros everywhere
-            # Put a portion of the counts into floor bin - the further away,the
-            # less counts:
-            # single_escape[i_floor] = (1-floor_distance) * f_pSE(E_fe)
-            # single_escape[i_ceil] = floor_distance * f_pSE(E_fe)
-            discrete_peaks[i_floor] += (1-floor_distance) * f_pSE(E_fe)
-            discrete_peaks[i_ceil] += floor_distance * f_pSE(E_fe)
-            # single_escape = gauss_smoothing(single_escape, Eout_array,
-                                            # fwhm_abs_array)  # Smoothe
-            # R[j, :] += single_escape
+                = self.two_channel_split(E_se, Eout)
+            discrete_peaks[
+                i_floor] += (1 - floor_distance) * self.f_pSE(E_fe)
+            discrete_peaks[i_ceil] += floor_distance * self.f_pSE(E_fe)
 
-        # Repeat for double-escape peak, at index i_de
-        E_de = E_fe - 2*511
-        if E_de >= 0 and E_de >= Eout_array[0]:
+        # Repeat for double-escape peak
+        E_de = E_fe - 2 * 511
+        if E_de >= 0 and E_de >= Eout[0]:
             i_floor, i_ceil, floor_distance\
-                = two_channel_split(E_de, Eout_array)
-            # double_escape = np.zeros(N_out)
-            # double_escape[i_floor] = (1-floor_distance) * f_pDE(E_fe)
-            # double_escape[i_ceil] = floor_distance * f_pDE(E_fe)
-            discrete_peaks[i_floor] += (1-floor_distance) * f_pDE(E_fe)
-            discrete_peaks[i_ceil] += floor_distance * f_pDE(E_fe)
-            # double_escape = gauss_smoothing(double_escape, Eout_array,
-                                            # fwhm_abs_array)  # Smoothe
-            # R[j, :] += double_escape
+                = self.two_channel_split(E_de, Eout)
+            discrete_peaks[
+                i_floor] += (1 - floor_distance) * self.f_pDE(E_fe)
+            discrete_peaks[i_ceil] += floor_distance * self.f_pDE(E_fe)
 
-        # Add 511 annihilation peak, at index i_an
-        if E_fe > 511 and 511 >= Eout_array[0]:
+        # Add 511 annihilation peak
+        if E_fe > 511 and 511 >= Eout[0]:
             E_511 = 511
             i_floor, i_ceil, floor_distance\
-                = two_channel_split(E_511, Eout_array)
-            # fiveeleven = np.zeros(N_out)
-            # fiveeleven[i_floor] = (1-floor_distance) * f_p511(E_fe)
-            # fiveeleven[i_ceil] = floor_distance * f_p511(E_fe)
-            discrete_peaks[i_floor] += (1-floor_distance) * f_p511(E_fe)
-            discrete_peaks[i_ceil] += floor_distance * f_p511(E_fe)
-            # fiveeleven = gauss_smoothing(fiveeleven, Eout_array,
-                                         # fwhm_abs_array)  # Smoothe
-            # R[j, :] += fiveeleven
+                = self.two_channel_split(E_511, Eout)
+            discrete_peaks[
+                i_floor] += (1 - floor_distance) * self.f_p511(E_fe)
+            discrete_peaks[i_ceil] += floor_distance * self.f_p511(E_fe)
 
-        # Do common smoothing of the discrete_peaks array:
-        discrete_peaks = gauss_smoothing(discrete_peaks, Eout_array,
-                                         fwhm_abs_array)  # Smoothe
+        if not self.smooth_compton:
+            # Do common smoothing of the discrete_peaks array:
+            discrete_peaks = gauss_smoothing(discrete_peaks, Eout,
+                                             fwhm_abs_array,
+                                             truncate=self.truncate)
+        return discrete_peaks
 
-        R[j, :] += discrete_peaks
+    @staticmethod
+    def E_compton(Eg, theta):
+        """
+        Calculates the energy of an electron that is scattered an angle
+        theta by a gamma-ray of energy Eg.
+        Adapted from MAMA, file "folding.f", which references
+        Canberra catalog ed.7, p.2.
 
-        # === Finally, normalise the row to unity (probability conservation): ===
-        R[j, :] = div0(R[j, :], np.sum(R[j, :]))
+        Note: For `Eg <= 0.1` it returns `Eg`. (workaround)
 
-    # END loop over Eout energies Ej
+        Args:
+            Eg: Energy of gamma-ray in keV
+            theta: Angle of scatter in radians
 
-    # Remove any negative elements from response matrix:
-    R[R < 0] = 0
+        Returns:
+            Energy Ee of scattered electron
+        """
+        scattered = Eg / (1 + Eg / 511 * (1 - np.cos(theta)))
+        electron = scattered * Eg / 511 * (1 - np.cos(theta))
+        return np.where(Eg > 0.1, electron, Eg)
 
-    # for i_plt in [j_test]:
-        # ax.plot(Eout_array, R[i_plt,:], label="interpolated, Eout = {:.0f}".format(Eout_array[i_plt]), linestyle="--")
+    @staticmethod
+    def dE_dtheta(Eg, theta):
+        """
+        Function to correct number of counts due to delta(theta)
+        Adapted from MAMA in the file kelvin.f
+        It is dE/dtheta of the E(theta) in Eq. (2) in Guttormsen 1996.
 
-    # ax.legend()
-    # plt.show()
+        Args:
+            Eg: Energy of gamma-ray in keV
+            theta: Angle of scatter in radians
 
-    response = Matrix(values=R, Eg=Eout_array, Ex=Eout_array)
+        Returns:
+            TYPE: dE_dtheta
+        """
+        a = (Eg * Eg / 511 * np.sin(theta))
+        b = (1 + Eg / 511 * (1 - np.cos(theta)))**2
+        return a / b
 
-    if return_table:
-        # Return the response matrix, as well as the other structures, FWHM and efficiency, interpolated to the Eout_array
-        response_table = {'E': Eout_array,
-                          'fwhm_abs': fwhm_abs_array,
-                          'fwhm_rel_%': f_fwhm_rel_perCent(Eout_array),
-                          'fwhm_rel': f_fwhm_rel_perCent(Eout_array)/100,
-                          'eff_tot': f_Eff_tot(Eout_array),
-                          'pcmp': f_pcmp(Eout_array),
-                          'pFE': f_pFE(Eout_array),
-                          'pSE': f_pSE(Eout_array),
-                          'pDE': f_pDE(Eout_array),
-                          'p511': f_p511(Eout_array)}
-        response_table = pd.DataFrame(data=response_table)
-        return response, response_table
-    else:
-        return response
+    @staticmethod
+    def two_channel_split(E_centroid, E_array):
+        """
+        When E_centroid is between two bins in E_array, this function
+        returns the indices of the two nearest bins and the distance to
+        the lower bin. The distance to the higher bin is 1-floor_distance
 
+        Args:
+            E_centroid (double): The energy of the centroid (mid-bin)
+            E_array (np.array, double): The energy grid to distribute
+        """
 
+        a0 = E_array[0]
+        a1 = E_array[1] - E_array[0]
 
-if __name__ == "__main__":
-    # folderpath = "oscar2017_scale1.0"
-    folderpath = "oscar2017_scale1.15"
-    Eout_array = np.linspace(200,1400,100)
-    FWHM = 300.0
-    print("response(\"{:s}\") =".format(folderpath), response(folderpath, Eout_array, FWHM))
+        bin_as_float = (E_centroid - a0) / a1
+        i_floor = int(np.floor(bin_as_float))
+        i_ceil = int(np.ceil(bin_as_float))
+        floor_distance = (bin_as_float - i_floor)
+
+        return i_floor, i_ceil, floor_distance
